feat: use multiple project types with shared functions
